Stricter region-exclusive test
We now test if a title was only released in one region, along with
"physical = yes" as part of the region-exclusive test.

There are a few tail cases that need to be considered, e.g.
region-exclusive in "Japan and US only" is treated as exclusive in NA
and JB - but I am still unsure how to define these cases.